Print chosen metric

Choose metric from parity loss metrics and plot it for every model. The one with the least parity loss is more fair in terms of this particular metric.

# S3 method for chosen_metric
print(x, ...)

Arguments

x
chosen_metric object

...
other chosen_metric object

Examples

data("german")

y_numeric <- as.numeric(german$Risk) -1

lm_model <- glm(Risk~.,
                data = german,
                family=binomial(link="logit"))

rf_model <- ranger::ranger(Risk ~.,
                           data = german,
                           probability = TRUE,
                           num.trees = 200)

explainer_lm <- DALEX::explain(lm_model, data = german[,-1], y = y_numeric)

#> Preparation of a new explainer is initiated
#>   -> model label       :  lm  (  default  )
#>   -> data              :  1000  rows  9  cols 
#>   -> target variable   :  1000  values 
#>   -> model_info        :  package stats , ver. 4.0.2 , task classification (  default  ) 
#>   -> predict function  :  yhat.glm  will be used (  default  )
#>   -> predicted values  :  numerical, min =  0.1369187 , mean =  0.7 , max =  0.9832426  
#>   -> residual function :  difference between y and yhat (  default  )
#>   -> residuals         :  numerical, min =  -0.9572803 , mean =  1.940006e-17 , max =  0.8283475  
#>   A new explainer has been created!  

explainer_rf <- DALEX::explain(rf_model, data = german[,-1], y = y_numeric)

#> Preparation of a new explainer is initiated
#>   -> model label       :  ranger  (  default  )
#>   -> data              :  1000  rows  9  cols 
#>   -> target variable   :  1000  values 
#>   -> model_info        :  package ranger , ver. 0.12.1 , task classification (  default  ) 
#>   -> predict function  :  yhat.ranger  will be used (  default  )
#>   -> predicted values  :  numerical, min =  0.08893254 , mean =  0.6962716 , max =  0.9970694  
#>   -> residual function :  difference between y and yhat (  default  )
#>   -> residuals         :  numerical, min =  -0.6978591 , mean =  0.003728397 , max =  0.6329794  
#>   A new explainer has been created!  

fobject <- fairness_check(explainer_lm, explainer_rf,
                          protected = german$Sex,
                          privileged = "male")

#> Creating fairness object
#> -> Privileged subgroup		: character ( Ok  )
#> -> Protected varaible		: factor ( Ok  ) 
#> -> Fairness objects		: compatible
#> -> Checking explainers		: compatible
#> -> Metric calculation		: successful
#>  Fairness object created succesfully  

cm <- choose_metric(fobject, "TPR_parity_loss")
print(cm)

#> 
#> choosen metric:
#>  TPR_parity_loss
#> data:
#>       metric  label
#> 1 0.08505568     lm
#> 2 0.01002004 ranger
#>