Print fairness PCA

Print principal components after using pca on fairness object

Print principal components after using pca on fairness object

# S3 method for fairness_pca
print(x, ...)

# S3 method for fairness_pca
print(x, ...)

Arguments

x
fairness object

...
other params

Value

Examples

library(DALEX)
library(ranger)

data("compas")
y_numeric <- as.numeric(compas$Two_yr_Recidivism) - 1
# models

rf_compas_1 <- ranger(Two_yr_Recidivism ~ Number_of_Priors + Age_Below_TwentyFive, data = compas, probability = TRUE)
rf_compas_3 <- ranger(Two_yr_Recidivism ~ Sex + Age_Above_FourtyFive, data = compas, probability = TRUE)
model_compas_lr <- glm(Two_yr_Recidivism ~ ., data = compas, family = binomial(link = "logit"))
rf_compas_5 <- ranger(Two_yr_Recidivism ~ ., data = compas, probability = TRUE)
rf_compas_6 <- ranger(Two_yr_Recidivism ~ Age_Above_FourtyFive + Misdemeanor, data = compas, probability = TRUE)
rf_compas_7 <- ranger(Two_yr_Recidivism ~ ., data = compas, probability = TRUE)
rf_compas_8 <- ranger(Two_yr_Recidivism ~ Sex + Age_Above_FourtyFive + Misdemeanor + Ethnicity, data = compas, probability = TRUE)

# explainers
explainer_1 <- explain(rf_compas_1, data = compas, y = y_numeric)

#> Preparation of a new explainer is initiated
#>   -> model label       :  ranger  (  default  )
#>   -> data              :  6172  rows  7  cols 
#>   -> target variable   :  6172  values 
#>   -> model_info        :  package ranger , ver. 0.12.1 , task classification (  default  ) 
#>   -> predict function  :  yhat.ranger  will be used (  default  )
#>   -> predicted values  :  numerical, min =  0.2973366 , mean =  0.4553661 , max =  0.767184  
#>   -> residual function :  difference between y and yhat (  default  )
#>   -> residuals         :  numerical, min =  -0.7399414 , mean =  -0.0002462405 , max =  0.7026634  
#>   A new explainer has been created!  

explainer_3 <- explain(rf_compas_3, data = compas, y = y_numeric)

#> Preparation of a new explainer is initiated
#>   -> model label       :  ranger  (  default  )
#>   -> data              :  6172  rows  7  cols 
#>   -> target variable   :  6172  values 
#>   -> model_info        :  package ranger , ver. 0.12.1 , task classification (  default  ) 
#>   -> predict function  :  yhat.ranger  will be used (  default  )
#>   -> predicted values  :  numerical, min =  0.2954243 , mean =  0.4550311 , max =  0.5018188  
#>   -> residual function :  difference between y and yhat (  default  )
#>   -> residuals         :  numerical, min =  -0.5018188 , mean =  8.874702e-05 , max =  0.7045757  
#>   A new explainer has been created!  

explainer_4 <- explain(model_compas_lr, data = compas, y = y_numeric)

#> Preparation of a new explainer is initiated
#>   -> model label       :  lm  (  default  )
#>   -> data              :  6172  rows  7  cols 
#>   -> target variable   :  6172  values 
#>   -> model_info        :  package stats , ver. 4.0.0 , task classification (  default  ) 
#>   -> predict function  :  yhat.glm  will be used (  default  )
#>   -> predicted values  :  numerical, min =  0.1144574 , mean =  0.4551199 , max =  0.995477  
#>   -> residual function :  difference between y and yhat (  default  )
#>   -> residuals         :  numerical, min =  -0.9767658 , mean =  5.050949e-13 , max =  0.8822826  
#>   A new explainer has been created!  

explainer_5 <- explain(rf_compas_5, data = compas, y = y_numeric)

#> Preparation of a new explainer is initiated
#>   -> model label       :  ranger  (  default  )
#>   -> data              :  6172  rows  7  cols 
#>   -> target variable   :  6172  values 
#>   -> model_info        :  package ranger , ver. 0.12.1 , task classification (  default  ) 
#>   -> predict function  :  yhat.ranger  will be used (  default  )
#>   -> predicted values  :  numerical, min =  0.1374646 , mean =  0.4549699 , max =  0.8544861  
#>   -> residual function :  difference between y and yhat (  default  )
#>   -> residuals         :  numerical, min =  -0.7832948 , mean =  0.0001499733 , max =  0.8364415  
#>   A new explainer has been created!  

explainer_6 <- explain(rf_compas_6, data = compas, y = y_numeric)

#> Preparation of a new explainer is initiated
#>   -> model label       :  ranger  (  default  )
#>   -> data              :  6172  rows  7  cols 
#>   -> target variable   :  6172  values 
#>   -> model_info        :  package ranger , ver. 0.12.1 , task classification (  default  ) 
#>   -> predict function  :  yhat.ranger  will be used (  default  )
#>   -> predicted values  :  numerical, min =  0.3052847 , mean =  0.4551928 , max =  0.5135196  
#>   -> residual function :  difference between y and yhat (  default  )
#>   -> residuals         :  numerical, min =  -0.5135196 , mean =  -7.287461e-05 , max =  0.6947153  
#>   A new explainer has been created!  

explainer_7 <- explain(rf_compas_7, data = compas, y = y_numeric)

#> Preparation of a new explainer is initiated
#>   -> model label       :  ranger  (  default  )
#>   -> data              :  6172  rows  7  cols 
#>   -> target variable   :  6172  values 
#>   -> model_info        :  package ranger , ver. 0.12.1 , task classification (  default  ) 
#>   -> predict function  :  yhat.ranger  will be used (  default  )
#>   -> predicted values  :  numerical, min =  0.1294032 , mean =  0.4549614 , max =  0.8387182  
#>   -> residual function :  difference between y and yhat (  default  )
#>   -> residuals         :  numerical, min =  -0.7863882 , mean =  0.0001585057 , max =  0.8461161  
#>   A new explainer has been created!  

# different labels
explainer_3$label <- "rf3"
explainer_4$label <- "glm"
explainer_5$label <- "rf5"
explainer_6$label <- "rf6"
explainer_7$label <- "rf7"
explainers <- list(explainer_1, explainer_3, explainer_4, explainer_5, explainer_6, explainer_7)

fobject <- create_fairness_object(explainer_1, explainer_3, explainer_4, explainer_5, explainer_6, explainer_7,
  outcome = "Two_yr_Recidivism",
  group = "Ethnicity",
  base = "Caucasian"
)

#> Getting data from first ( ranger )  explainer 

#> 
#> 0's appear in confusion matrix for group: Native_American

#> 
#> 0's appear in confusion matrix for group: Asian

#> NA's created in metric's matrix 

fpca <- pca(fobject)

#> Ommiting metrics with NA:  F1_parity_loss

#> 

print(fpca)

#> Fairness PCA : 
#>             PC1        PC2         PC3         PC4          PC5           PC6
#> [1,] -0.8664882  0.3604406 -0.10292139 -0.38247389 -0.205944717  6.938894e-17
#> [2,]  4.7193638 -0.2905295 -0.08625036  0.03062556 -0.009221070  1.110223e-16
#> [3,] -1.1891817 -3.9454526  0.58808851  0.11046711  0.007128958  3.608225e-16
#> [4,] -1.2685847  0.5206465 -1.30233701  0.30562255 -0.031103621 -1.110223e-16
#> [5,] -0.7010494  2.7821939  1.12830492  0.20831445 -0.009542101  1.665335e-16
#> [6,] -0.6940597  0.5727011 -0.22488467 -0.27255578  0.248682551  9.159340e-16
#> 
#> Created with: 
#> [1] "ranger" "rf3"    "glm"    "rf5"    "rf6"    "rf7"   
#> 
#> First two components explained 93 % of variance.

library(DALEX)
library(ranger)

data("compas")
y_numeric <- as.numeric(compas$Two_yr_Recidivism) - 1
# models

rf_compas_1 <- ranger(Two_yr_Recidivism ~ Number_of_Priors + Age_Below_TwentyFive, data = compas, probability = TRUE)
rf_compas_3 <- ranger(Two_yr_Recidivism ~ Sex + Age_Above_FourtyFive, data = compas, probability = TRUE)
model_compas_lr <- glm(Two_yr_Recidivism ~ ., data = compas, family = binomial(link = "logit"))
rf_compas_5 <- ranger(Two_yr_Recidivism ~ ., data = compas, probability = TRUE)
rf_compas_6 <- ranger(Two_yr_Recidivism ~ Age_Above_FourtyFive + Misdemeanor, data = compas, probability = TRUE)
rf_compas_7 <- ranger(Two_yr_Recidivism ~ ., data = compas, probability = TRUE)
rf_compas_8 <- ranger(Two_yr_Recidivism ~ Sex + Age_Above_FourtyFive + Misdemeanor + Ethnicity, data = compas, probability = TRUE)

# explainers
explainer_1 <- explain(rf_compas_1, data = compas, y = y_numeric)

#> Preparation of a new explainer is initiated
#>   -> model label       :  ranger  (  default  )
#>   -> data              :  6172  rows  7  cols 
#>   -> target variable   :  6172  values 
#>   -> model_info        :  package ranger , ver. 0.12.1 , task classification (  default  ) 
#>   -> predict function  :  yhat.ranger  will be used (  default  )
#>   -> predicted values  :  numerical, min =  0.297975 , mean =  0.4551935 , max =  0.7710185  
#>   -> residual function :  difference between y and yhat (  default  )
#>   -> residuals         :  numerical, min =  -0.7515241 , mean =  -7.362045e-05 , max =  0.702025  
#>   A new explainer has been created!  

explainer_3 <- explain(rf_compas_3, data = compas, y = y_numeric)

#> Preparation of a new explainer is initiated
#>   -> model label       :  ranger  (  default  )
#>   -> data              :  6172  rows  7  cols 
#>   -> target variable   :  6172  values 
#>   -> model_info        :  package ranger , ver. 0.12.1 , task classification (  default  ) 
#>   -> predict function  :  yhat.ranger  will be used (  default  )
#>   -> predicted values  :  numerical, min =  0.2927357 , mean =  0.4557895 , max =  0.5007217  
#>   -> residual function :  difference between y and yhat (  default  )
#>   -> residuals         :  numerical, min =  -0.5007217 , mean =  -0.0006695689 , max =  0.7072643  
#>   A new explainer has been created!  

explainer_4 <- explain(model_compas_lr, data = compas, y = y_numeric)

#> Preparation of a new explainer is initiated
#>   -> model label       :  lm  (  default  )
#>   -> data              :  6172  rows  7  cols 
#>   -> target variable   :  6172  values 
#>   -> model_info        :  package stats , ver. 4.0.0 , task classification (  default  ) 
#>   -> predict function  :  yhat.glm  will be used (  default  )
#>   -> predicted values  :  numerical, min =  0.1144574 , mean =  0.4551199 , max =  0.995477  
#>   -> residual function :  difference between y and yhat (  default  )
#>   -> residuals         :  numerical, min =  -0.9767658 , mean =  5.050949e-13 , max =  0.8822826  
#>   A new explainer has been created!  

explainer_5 <- explain(rf_compas_5, data = compas, y = y_numeric)

#> Preparation of a new explainer is initiated
#>   -> model label       :  ranger  (  default  )
#>   -> data              :  6172  rows  7  cols 
#>   -> target variable   :  6172  values 
#>   -> model_info        :  package ranger , ver. 0.12.1 , task classification (  default  ) 
#>   -> predict function  :  yhat.ranger  will be used (  default  )
#>   -> predicted values  :  numerical, min =  0.1284616 , mean =  0.4554075 , max =  0.8530789  
#>   -> residual function :  difference between y and yhat (  default  )
#>   -> residuals         :  numerical, min =  -0.7923859 , mean =  -0.0002875591 , max =  0.8442875  
#>   A new explainer has been created!  

explainer_6 <- explain(rf_compas_6, data = compas, y = y_numeric)

#> Preparation of a new explainer is initiated
#>   -> model label       :  ranger  (  default  )
#>   -> data              :  6172  rows  7  cols 
#>   -> target variable   :  6172  values 
#>   -> model_info        :  package ranger , ver. 0.12.1 , task classification (  default  ) 
#>   -> predict function  :  yhat.ranger  will be used (  default  )
#>   -> predicted values  :  numerical, min =  0.3049465 , mean =  0.4545463 , max =  0.5141195  
#>   -> residual function :  difference between y and yhat (  default  )
#>   -> residuals         :  numerical, min =  -0.5141195 , mean =  0.0005735709 , max =  0.6950535  
#>   A new explainer has been created!  

explainer_7 <- explain(rf_compas_7, data = compas, y = y_numeric)

#> Preparation of a new explainer is initiated
#>   -> model label       :  ranger  (  default  )
#>   -> data              :  6172  rows  7  cols 
#>   -> target variable   :  6172  values 
#>   -> model_info        :  package ranger , ver. 0.12.1 , task classification (  default  ) 
#>   -> predict function  :  yhat.ranger  will be used (  default  )
#>   -> predicted values  :  numerical, min =  0.1352772 , mean =  0.455271 , max =  0.8493591  
#>   -> residual function :  difference between y and yhat (  default  )
#>   -> residuals         :  numerical, min =  -0.7902227 , mean =  -0.0001511537 , max =  0.8406991  
#>   A new explainer has been created!  

# different labels
explainer_3$label <- "rf3"
explainer_4$label <- "glm"
explainer_5$label <- "rf5"
explainer_6$label <- "rf6"
explainer_7$label <- "rf7"
explainers <- list(explainer_1, explainer_3, explainer_4, explainer_5, explainer_6, explainer_7)

fobject <- create_fairness_object(explainer_1, explainer_3, explainer_4, explainer_5, explainer_6, explainer_7,
  outcome = "Two_yr_Recidivism",
  group = "Ethnicity",
  base = "Caucasian"
)

#> Getting data from first ( ranger )  explainer 

#> 
#> 0's appear in confusion matrix for group: Native_American

#> 
#> 0's appear in confusion matrix for group: Asian

#> NA's created in metric's matrix 

fpca <- pca(fobject)

#> Ommiting metrics with NA:  F1_parity_loss

#> 

print(fpca)

#> Fairness PCA : 
#>              PC1         PC2         PC3          PC4           PC5
#> [1,] -1.05385678  0.08901173  0.37080474  0.336588465 -3.776064e-03
#> [2,]  4.60546882  1.03422810 -0.04527607  0.009008656  2.589179e-05
#> [3,] -0.02763397 -4.30733466 -0.47731446 -0.029790145  1.256751e-04
#> [4,] -0.91293000  0.33597028  0.58996190 -0.166548259 -4.071330e-02
#> [5,] -1.66676167  2.48059114 -0.98402358 -0.026871306 -8.359936e-04
#> [6,] -0.94428640  0.36753341  0.54584748 -0.122387412  4.517379e-02
#>                PC6
#> [1,]  1.734723e-16
#> [2,]  6.383782e-16
#> [3,]  1.656661e-16
#> [4,]  2.428613e-16
#> [5,] -1.942890e-16
#> [6,]  1.734723e-16
#> 
#> Created with: 
#> [1] "ranger" "rf3"    "glm"    "rf5"    "rf6"    "rf7"   
#> 
#> First two components explained 96 % of variance.